requirements.txt
# Основные зависимости для Telegram-бота
python-telegram-bot==20.7
requests==2.31.0

# Работа с базой данных
peewee==3.16.3
#playhouse==3.16.3  # Часть peewee, но иногда требуется отдельно

# Дополнительные утилиты
python-dotenv==1.0.0  # Для работы с переменными окружения
cryptography==41.0.7  # Для шифрования конфигураций
